seperate wrapped closer struct to wrapped_closer.go
